backend/testing/requirements.txt
google-generativeai==0.6.0
SpeechRecognition==3.10.4
PyAudio==0.2.14
elevenlabs==2.22.0
playsound==1.2.2
packaging==25.0

